move chat elements in chat folder
