+++ static/js/elements/chat/msg.js
